% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecuador_Species_List.R
\docType{data}
\name{Ecuador_Species_List}
\alias{Ecuador_Species_List}
\title{Dataset helpfile header . . .}
\format{
A data frame with 87 rows and 10 column(s)
\describe{
\item{Species}{Describe column Species here . . .}
\item{Specie.Code}{Describe column Specie.Code here . . .}
\item{Body_Size}{Describe column Body_Size here . . .}
\item{Diet}{Describe column Diet here . . .}
\item{Foraging_Strata}{Describe column Foraging_Strata here . . .}
\item{Primary_Habitat}{Describe column Primary_Habitat here . . .}
\item{Habitat_Breadth}{Describe column Habitat_Breadth here . . .}
\item{Rarity_Range_Size}{Describe column Rarity_Range_Size here . . .}
\item{Rarity_Local_Abund}{Describe column Rarity_Local_Abund here . . .}
\item{Rarity_Neotrop_Abund}{Describe column Rarity_Neotrop_Abund here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
Ecuador_Species_List
}
\description{
Short description of data . . .
}
\keyword{datasets}
